msm: mdss: compile mdss drivers with arm 64-bit toolchain
Fix mdss drivers to compile with arm 64-bit tool chain
by addressing incorrect typecasts and include appropriate
headers.

+++ drivers/video/msm/mdss/mdss_mdp_pp.c
@@ -5228,7 +5228,7 @@ int mdss_mdp_calib_config(struct mdp_calib_config_data *cfg, u32 *copyback)
 	void *ptr;
 
 	/* Calib addrs are always offsets from the MDSS base */
-	ptr = (void *)((unsigned int) cfg->addr) +
+	ptr = (void *)((unsigned long) cfg->addr) +
 		((uintptr_t) mdss_res->mdss_base);
 	if (is_valid_calib_addr(ptr, cfg->ops))
 		ret = 0;
